pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "octofiestaspotifyapi"
version = "0.1.0"
description = "Spotify API for Octo-Fiesta"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "httpx>=0.27",
  "injector>=0.22",
  "pydantic>=2.9",
  "python-dotenv>=1.0",
  "uvicorn>=0.30",
  "camoufox>=0.3",
  "playwright>=1.48",
]

[tool.setuptools]

[tool.setuptools.packages.find]
include = ["octofiestaspotifyapi*"]
